feat: Add support for custom A/AAAA addresses
With the addition of addresses variable the module can make DNS zone
records that resolve to private addresses other than those for
`restricted.googleapis.com` or `private.googleapis.com`. The reason for
this change is to support PSC option for Google APIs access.

Closes #74
BREAKING CHANGE: The module has changed minimal requirements for
terraform/tofu version to 1.5+ and Google provider to 6.0+.